LayoutTests/platform/wincairo/accessibility/w3c-svg-description-calculation-expected.txt

This tests the accessible text alternatives results for SVG.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


test1:
	Expected description: end
	AXTitle:
	AXDescription: abc
test2:
	Expected description: end
	AXTitle:
	AXDescription:
test3:
	Expected description: end
	AXTitle:
	AXDescription: abc
test4:
	Expected description: end
	AXTitle:
	AXDescription:
test5:
	Expected description: 123
	AXTitle:
	AXDescription: abc
test6:
	Expected description: 123
	AXTitle:
	AXDescription:
test7:
	Expected description: end
	AXTitle:
	AXDescription: abc
test8:
	Expected description: 554
	AXTitle:
	AXDescription: abc
test9:
	Expected description: 567
	AXTitle:
	AXDescription:
test10:
	Expected description: 123
	AXTitle:
	AXDescription:
test11:
	Expected description: 123
	AXTitle:
	AXDescription:
test12:
	Expected description: 123
	AXTitle:
	AXDescription: abc
test13:
	Expected description: 12
	AXTitle:
	AXDescription:
test14:
	Expected description: (empty)
	AXTitle:
	AXDescription: abc
test15:
	Expected description: (empty)
	AXTitle:
	AXDescription:
test16:
	Expected description: 222
	AXTitle:
	AXDescription: abc
test17:
	Expected description: 911
	AXTitle:
	AXDescription:
test18:
	Expected description: 123
	AXTitle:
	AXDescription: abc
test19:
	Expected description: 123
	AXTitle:
	AXDescription:
test20:
	Expected description: the end
	AXTitle:
	AXDescription:
test21:
	Expected description: the end
	AXTitle:
	AXDescription:
test22:
	Expected description: end
	AXTitle:
	AXDescription:
test23:
	Expected description: abc
	AXTitle: end
	AXDescription: end
test24:
	Expected description: abc
	AXTitle:
	AXDescription: bob
test25:
	Expected description: counter
	AXTitle: end
	AXDescription: end
test26:
	Expected description: counter
	AXTitle:
	AXDescription: bob
test27:
	Expected description: barra
	AXTitle: end
	AXDescription: end
test28:
	Expected description: barra
	AXTitle:
	AXDescription: bob
test29:
	Expected description: (empty)
	AXTitle: abc
	AXDescription: abc
test30:
	Expected description: (empty)
	AXTitle:
	AXDescription: bob
test31:
	Expected description: (empty)
	AXTitle: abc
	AXDescription: abc
test32:
	Expected description: (empty)
	AXTitle:
	AXDescription: bob
test33:
	Expected description: (empty)
	AXTitle:
	AXDescription: bob
test34:
	Expected description: (empty)
	AXTitle:
	AXDescription: end
test35:
	Expected description: (empty)
	AXTitle:
	AXDescription: end
test36:
	Expected description: (empty)
	AXTitle:
	AXDescription: booth
test37:
	Expected description: (empty)
	AXTitle:
	AXDescription: counter
test38:
	Expected description: (empty)
	AXTitle:
	AXDescription: table
test39:
	Expected description: (empty)
	AXTitle:
	AXDescription:
test40:
	Expected description: (empty)
	AXTitle:
	AXDescription:
test41:
	Expected description: (empty)
	AXTitle:
	AXDescription:
test42:
	Expected description: (empty)
	AXTitle:
	AXDescription:
test43:
	Expected description: (empty)
	AXTitle:
	AXDescription:
PASS successfullyParsed is true

TEST COMPLETE
